+++ source/lib/report-network-problem.js
@@ -1,7 +1,7 @@
 // @flow
 
-import {tracker} from '../analytics'
-import bugsnag from '../bugsnag'
+import {tracker} from '../init/analytics'
+import bugsnag from '../init/bugsnag'
 
 export function reportNetworkProblem(err: Error) {
 	tracker.trackException(err.message)
