+++ source/redux/parts/settings.js
@@ -14,7 +14,7 @@ import {
 	setAcknowledgementStatus,
 } from '../../lib/storage'
 
-import {trackLogOut, trackLogIn, trackLoginFailure} from '../../analytics'
+import {trackLogOut, trackLogIn, trackLoginFailure} from '../../init/analytics'
 
 import {type ReduxState} from '../index'
 import {type UpdateBalancesType, updateBalances} from './balances'
