+++ source/root.js
@@ -5,7 +5,7 @@ import App from './app'
 
 // I'm not importing the exported variable because I just want to initialize
 // the file here.
-import './bugsnag'
+import './init/bugsnag'
 
 YellowBox.ignoreWarnings([
 	// TODO: remove me after upgrading to RN 0.56
