+++ source/views/calendar/event-list.js
@@ -4,7 +4,7 @@ import * as React from 'react'
 import {StyleSheet, SectionList} from 'react-native'
 import * as c from '../../components/colors'
 import toPairs from 'lodash/toPairs'
-import {trackCalendarEventOpen} from '../../analytics'
+import {trackCalendarEventOpen} from '../../init/analytics'
 import type {TopLevelViewPropsType} from '../types'
 import type {EventType, PoweredBy} from './types'
 import groupBy from 'lodash/groupBy'
