+++ source/views/contacts/detail.js
@@ -7,11 +7,11 @@ import {Markdown} from '../../components/markdown'
 import {ListFooter} from '../../components/list'
 import glamorous from 'glamorous-native'
 import {callPhone} from '../../components/call-phone'
-import {tracker} from '../../analytics'
+import {tracker} from '../../init/analytics'
 import {Button} from '../../components/button'
 import {openUrl} from '../../components/open-url'
 import type {ContactType} from './types'
-import {GH_NEW_ISSUE_URL} from '../../globals'
+import {GH_NEW_ISSUE_URL} from '../../init/constants'
 
 const Title = glamorous.text({
 	fontSize: 36,
