+++ source/views/menus/menu-bonapp.js
@@ -23,11 +23,11 @@ import {trimStationName, trimItemLabel} from './lib/trim-names'
 import {getTrimmedTextWithSpaces, parseHtml} from '../../lib/html'
 import {AllHtmlEntities} from 'html-entities'
 import {toLaxTitleCase} from 'titlecase'
-import {tracker} from '../../analytics'
-import bugsnag from '../../bugsnag'
+import {tracker} from '../../init/analytics'
+import bugsnag from '../../init/bugsnag'
 import delay from 'delay'
 import retry from 'p-retry'
-import {API} from '../../globals'
+import {API} from '../../init/constants'
 
 const CENTRAL_TZ = 'America/Winnipeg'
 const entities = new AllHtmlEntities()
