+++ source/views/sis/student-work/index.js
@@ -6,8 +6,8 @@ import {TabBarIcon} from '../../../components/tabbar-icon'
 import type {TopLevelViewPropsType} from '../../types'
 import * as c from '../../../components/colors'
 import {ListSeparator, ListSectionHeader} from '../../../components/list'
-import {tracker, trackStudentJobOpen} from '../../../analytics'
-import bugsnag from '../../../bugsnag'
+import {tracker, trackStudentJobOpen} from '../../../init/analytics'
+import bugsnag from '../../../init/bugsnag'
 import {NoticeView} from '../../../components/notice'
 import LoadingView from '../../../components/loading'
 import delay from 'delay'
@@ -16,7 +16,7 @@ import orderBy from 'lodash/orderBy'
 import groupBy from 'lodash/groupBy'
 import {toLaxTitleCase as titleCase} from 'titlecase'
 import {JobRow} from './job-row'
-import {API} from '../../../globals'
+import {API} from '../../../init/constants'
 import type {JobType} from './types'
 
 const jobsUrl = API('/jobs')
