+++ source/views/transportation/other-modes/list.js
@@ -13,7 +13,7 @@ import groupBy from 'lodash/groupBy'
 import toPairs from 'lodash/toPairs'
 import type {TopLevelViewPropsType} from '../../types'
 import type {OtherModeType} from '../types'
-import {API} from '../../../globals'
+import {API} from '../../../init/constants'
 
 const transportationUrl = API('/transit/modes')
